+++ src/collective/volto/formsupport/restapi/services/form_data/form_data.py
@@ -4,6 +4,7 @@ from collective.volto.formsupport.interfaces import IFormDataStore
 from collective.volto.formsupport.utils import get_blocks
 from plone import api
 from plone.memoize import view
+from plone.namedfile import NamedBlobFile
 from plone.restapi.interfaces import IExpandableElement
 from plone.restapi.serializer.converters import json_compatible
 from plone.restapi.services import Service
@@ -49,7 +50,6 @@ class FormData(object):
     @view.memoize
     def form_block(self):
         blocks = get_blocks(self.context)
-
         if isinstance(blocks, six.text_type):
             blocks = json.loads(blocks)
         if not blocks:
@@ -70,10 +70,21 @@ class FormData(object):
         for k, v in record.attrs.items():
             if k in ["fields_labels", "fields_order"]:
                 continue
-            data[k] = {
-                "value": json_compatible(v),
-                "label": fields_labels.get(k, k),
-            }
+            if isinstance(v, NamedBlobFile):
+                data[k] = {
+                    "value": {
+                        "url": f"{self.context.absolute_url()}/saved_data/@@download/{record.intid}/{k}/{v.filename}",
+                        "filename": v.filename,
+                        "contentType": v.contentType,
+                        "size": v.getSize(),
+                    },
+                    "label": fields_labels.get(k, k),
+                }
+            else:
+                data[k] = {
+                    "value": json_compatible(v),
+                    "label": fields_labels.get(k, k),
+                }
         data["id"] = record.intid
         return data