+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 2.7.1 (unreleased)
 ------------------
 
+- Save attachments as blobfile in the storage adapater, add a view to download them.
+  [mamico]
 - Allow attaching an XML version of the form data to the sent email #22
   [JeffersonBledsoe]
 - Allow the IDs of fields to be customised for CSV download and XML attaachments #22